Extraida la clase placeholder y refactorizada
