+++ app/src/main/java/com/llevame_app_project/PlaceholderFragment.java
@@ -1,0 +1,28 @@
+package com.llevame_app_project;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import com.llevame_app_project.R;
+
+/**
+ * A placeholder fragment containing a simple view.
+ */
+public class PlaceholderFragment extends Fragment {
+
+    public PlaceholderFragment() {
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.fragment_passenger, container, false);
+        TextView textView = (TextView) rootView.findViewById(R.id.section_label);
+        textView.setText("Bla");
+        return rootView;
+    }
+}
